+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # changelog
 
+ * 2.3.0 _May.30.2023_
+   * [add initial support](https://github.com/iambumblehead/esmock/pull/205) for the solution to "globalThis" mocks,
+   * support injecting definitions into the mock import tree,
+   * demonstrate mock setTimeout, fetch and Date scenarious at unit-tests
  * 2.2.3 _May.16.2023_
    * [add node v21 nightly](https://github.com/iambumblehead/esmock/pull/199) to test ci pipeline
    * send wide uri definitions to loader using loader worker